MyBOT_RAG/requirements.txt
openai 
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
faiss-cpu
pandas
python-dotenv
streamlit
transformers
torch 